backend/requirements.txt
# Django and REST Framework
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Database and utilities
python-dotenv==1.2.1
pillow==12.0.0

# Machine Learning and Recommendation Engine
scikit-learn==1.7.2
faiss-cpu==1.13.1
numpy==2.2.6
scipy==1.15.3
joblib==1.5.2
